refactor[core,manage,auth]: change `Cats` prefix
