+++ auth/src/main/scala/janstenpickle/vault/auth/UserPass.scala
@@ -1,9 +1,9 @@
 package janstenpickle.vault.auth
 
 import io.circe.generic.auto._
-import janstenpickle.scala.syntax.CatsAsyncResult._
-import janstenpickle.scala.syntax.CatsRequest._
-import janstenpickle.scala.syntax.CatsResponse._
+import janstenpickle.scala.syntax.SyntaxAsyncResult._
+import janstenpickle.scala.syntax.SyntaxRequest._
+import janstenpickle.scala.syntax.SyntaxResponse._
 import janstenpickle.vault.core.WSClient
 import uscala.concurrent.result.AsyncResult
